fix warning

+++ Source/Tcp/TcpSelect/TcpSelectBase.h
@@ -37,7 +37,7 @@ protected:
 protected:
 	fd_set m_RecvFds;
 	fd_set m_SendFds;
-	int m_MaxID;
+	uint64_t m_MaxID;
 	
 	std::map<int, std::list<TcpEvent*>> m_SendEvents;
 	std::list<int> m_DisConnectSessions;
